+++ forui/CHANGELOG.md
@@ -15,6 +15,9 @@ This update focuses on overhauling the API to better support the new dot-shortha
 * **Breaking** Remove `FAutocompleteStyle.popoverStyle` - use `FAutocompleteStyle.contentStyle` instead.
 
 ### `FBottomNavigationBar`
+* Add `FBottomNavigationBar.safeAreaTop`.
+* Add `FBottomNavigationBar.safeAreaBottom`.
+
 * Change `FBottomNavigationBarItem.label` to be optional.
 
 
@@ -26,7 +29,7 @@ This update focuses on overhauling the API to better support the new dot-shortha
 * **Breaking** Remove `FCheckbox.grouped(...)`. Use `FSelectGroupItemMixin.checkbox(...)` instead.
 
 
-### `FDateField`
+### `FDateField` 
 *  **Breaking** Change `FDateField.shift` to `FDateField.overflow`.
 
 
@@ -95,8 +98,9 @@ This update focuses on overhauling the API to better support the new dot-shortha
 * Add `FSliderMark.mark(...)`.
 
 
-### `FTab`
+### `FTabs`
 * Add `FTabEntry.entry(...)`.
+* Add `FTabs.mouseCursor`.
 
 
 ### `FTappable`
@@ -273,6 +277,8 @@ We've reworked `FProgress` to be more customizable and easier to use.
 
 ### `FTab`
 * Add `FTabMotion`.
+* Add `FTabs.mouseCursor`.
+
 * **Breaking** Replace `FTabController(animationDuration: ...)` with `FTabController(motion: ...)`.